requirements.txt
torch
torchvision
fastapi
uvicorn
Pillow
python-multipart